Renamed directory and files from submesh to subset
