+++ noether_filtering/include/noether_filtering/subset_extraction/extruded_polygon_subset_extractor.h
@@ -15,7 +15,7 @@
  */
 #pragma once
 
-#include <noether_filtering/submesh_extraction/submesh_extractor.h>
+#include <noether_filtering/subset_extraction/subset_extractor.h>
 
 namespace noether
 {
